paper-input is now part of selectable array
Shift the index when selecting a result because paper-input correspond
to index 0.
Using mouseenter instead of mouseover for to hover the span to do not
trigger the event.
When the mouse leave the result array, selection is reseted to
paper-input.
Take care to do not color the paper-input if selected.