Fix typeahead for polymer 2.
Because is= isn't supported in polymer2, the polymer 2 version of
paper-typeahead's input was just a regular input, and not an iron-input.
This resulted in bugs due to iron-input functions not getting triggered.

This is fixed the same way polymerElements made paper-input into a
hybrid element- by swapping out different versions of the iron-input
element based on which version of polymer is being used.